+++ scipy/special/tests/test_mpmath.py
@@ -4,18 +4,20 @@ Test Scipy functions versus mpmath, if available.
 """
 from __future__ import division, print_function, absolute_import
 
-import sys
-import time
-
 import numpy as np
 from numpy.testing import dec, run_module_suite, assert_, assert_allclose
 from numpy import pi
 
 import scipy.special as sc
-from scipy._lib.six import reraise, with_metaclass
+from scipy._lib.six import with_metaclass
 from scipy._lib._testutils import knownfailure_overridable
 from scipy.special._testutils import (MissingModule, check_version, FuncData,
-                                      assert_func_equal, DecoratorMeta)
+                                      DecoratorMeta, assert_func_equal)
+from scipy.special._mptestutils import (Arg, FixedArg, ComplexArg, IntArg,
+                                        assert_mpmath_equal,
+                                        nonfunctional_tooslow, trace_args,
+                                        time_limited, exception_to_nan,
+                                        inf_to_nan)
 from scipy.special._ufuncs import _sinpi, _cospi
 
 try:
@@ -355,6 +357,7 @@ def test_loggamma_taylor2():
 # ------------------------------------------------------------------------------
 
 @check_version(mpmath, '0.19')
+@dec.slow
 def test_rgamma_zeros():
     # Test around the zeros at z = 0, -1, -2, ...,  -169. (After -169 we
     # get values that are out of floating point range even when we're
@@ -381,6 +384,7 @@ def test_rgamma_zeros():
 # ------------------------------------------------------------------------------
 
 @check_version(mpmath, '0.19')
+@dec.slow
 def test_digamma_roots():
     # Test the special-cased roots for digamma.
     root = mpmath.findroot(mpmath.digamma, 1.5)
@@ -410,7 +414,7 @@ def test_digamma_negreal():
     # TestSystematic because the points need some jiggering so that
     # mpmath doesn't take forever.
 
-    digamma = _exception_to_nan(mpmath.digamma)
+    digamma = exception_to_nan(mpmath.digamma)
 
     x = -np.logspace(300, -30, 100)
     y = np.r_[-np.logspace(0, -3, 5), 0, np.logspace(-3, 0, 5)]
@@ -452,6 +456,7 @@ def test_digamma_boundary():
 # ------------------------------------------------------------------------------
 
 @check_version(mpmath, '0.19')
+@dec.slow
 def test_gammainc_boundary():
     # Test the transition to the asymptotic series.
     small = 25
@@ -473,6 +478,7 @@ def test_gammainc_boundary():
 # ------------------------------------------------------------------------------
 
 @check_version(mpmath, '0.19')
+@dec.slow
 def test_spence_circle():
     # The trickiest region for spence is around the circle |z - 1| = 1,
     # so test that region carefully.
@@ -481,7 +487,7 @@ def test_spence_circle():
         return complex(mpmath.polylog(2, 1 - z))
 
     r = np.linspace(0.5, 1.5)
-    theta = np.linspace(0, 2*np.pi)
+    theta = np.linspace(0, 2*pi)
     z = (1 + np.outer(r, np.exp(1j*theta))).flatten()
     dataset = []
     for z0 in z:
@@ -527,321 +533,6 @@ def test_cospi_zeros():
 
     dataset = np.array(dataset)
     FuncData(_cospi, dataset, 0, 1, rtol=2*eps).check()
-
-
-# ------------------------------------------------------------------------------
-# Machinery for systematic tests
-# ------------------------------------------------------------------------------
-
-class Arg(object):
-    """
-    Generate a set of numbers on the real axis, concentrating on
-    'interesting' regions and covering all orders of magnitude.
-    """
-
-    def __init__(self, a=-np.inf, b=np.inf, inclusive_a=True, inclusive_b=True):
-        self.a = a
-        self.b = b
-        self.inclusive_a = inclusive_a
-        self.inclusive_b = inclusive_b
-        if self.a == -np.inf:
-            self.a = -np.finfo(float).max/2
-        if self.b == np.inf:
-            self.b = np.finfo(float).max/2
-
-    def values(self, n):
-        """Return an array containing approximatively `n` numbers."""
-        n1 = max(2, int(0.3*n))
-        n2 = max(2, int(0.2*n))
-        n3 = max(8, n - n1 - n2)
-
-        v1 = np.linspace(-1, 1, n1)
-        v2 = np.r_[np.linspace(-10, 10, max(0, n2-4)),
-                   -9, -5.5, 5.5, 9]
-        if self.a >= 0 and self.b > 0:
-            v3 = np.r_[
-                np.logspace(-30, -1, 2 + n3//4),
-                np.logspace(5, np.log10(self.b), 1 + n3//4),
-                ]
-            v4 = np.logspace(1, 5, 1 + n3//2)
-        elif self.a < 0 < self.b:
-            v3 = np.r_[
-                np.logspace(-30, -1, 2 + n3//8),
-                np.logspace(5, np.log10(self.b), 1 + n3//8),
-                -np.logspace(-30, -1, 2 + n3//8),
-                -np.logspace(5, np.log10(-self.a), 1 + n3//8)
-                ]
-            v4 = np.r_[
-                np.logspace(1, 5, 1 + n3//4),
-                -np.logspace(1, 5, 1 + n3//4)
-                ]
-        elif self.b < 0:
-            v3 = np.r_[
-                -np.logspace(-30, -1, 2 + n3//4),
-                -np.logspace(5, np.log10(-self.b), 1 + n3//4),
-                ]
-            v4 = -np.logspace(1, 5, 1 + n3//2)
-        else:
-            v3 = []
-            v4 = []
-        v = np.r_[v1, v2, v3, v4, 0]
-        if self.inclusive_a:
-            v = v[v >= self.a]
-        else:
-            v = v[v > self.a]
-        if self.inclusive_b:
-            v = v[v <= self.b]
-        else:
-            v = v[v < self.b]
-        return np.unique(v)
-
-
-class FixedArg(object):
-    def __init__(self, values):
-        self._values = np.asarray(values)
-
-    def values(self, n):
-        return self._values
-
-
-class ComplexArg(object):
-    def __init__(self, a=complex(-np.inf, -np.inf), b=complex(np.inf, np.inf)):
-        self.real = Arg(a.real, b.real)
-        self.imag = Arg(a.imag, b.imag)
-
-    def values(self, n):
-        m = max(2, int(np.sqrt(n)))
-        x = self.real.values(m)
-        y = self.imag.values(m)
-        return (x[:,None] + 1j*y[None,:]).ravel()
-
-
-class IntArg(object):
-    def __init__(self, a=-1000, b=1000):
-        self.a = a
-        self.b = b
-
-    def values(self, n):
-        v1 = Arg(self.a, self.b).values(max(1 + n//2, n-5)).astype(int)
-        v2 = np.arange(-5, 5)
-        v = np.unique(np.r_[v1, v2])
-        v = v[(v >= self.a) & (v < self.b)]
-        return v
-
-
-class MpmathData(object):
-    def __init__(self, scipy_func, mpmath_func, arg_spec, name=None,
-                 dps=None, prec=None, n=5000, rtol=1e-7, atol=1e-300,
-                 ignore_inf_sign=False, nan_ok=True, param_filter=None):
-        self.scipy_func = scipy_func
-        self.mpmath_func = mpmath_func
-        self.arg_spec = arg_spec
-        self.dps = dps
-        self.prec = prec
-        self.n = n
-        self.rtol = rtol
-        self.atol = atol
-        self.ignore_inf_sign = ignore_inf_sign
-        self.nan_ok = nan_ok
-        if isinstance(self.arg_spec, np.ndarray):
-            self.is_complex = np.issubdtype(self.arg_spec.dtype, np.complexfloating)
-        else:
-            self.is_complex = any([isinstance(arg, ComplexArg) for arg in self.arg_spec])
-        self.ignore_inf_sign = ignore_inf_sign
-        if not name or name == '<lambda>':
-            name = getattr(scipy_func, '__name__', None)
-        if not name or name == '<lambda>':
-            name = getattr(mpmath_func, '__name__', None)
-        self.name = name
-        self.param_filter = param_filter
-
-    def check(self):
-        np.random.seed(1234)
-
-        # Generate values for the arguments
-        if isinstance(self.arg_spec, np.ndarray):
-            argarr = self.arg_spec.copy()
-        else:
-            num_args = len(self.arg_spec)
-            ms = np.asarray([1.5 if isinstance(arg, ComplexArg) else 1.0
-                             for arg in self.arg_spec])
-            ms = (self.n**(ms/sum(ms))).astype(int) + 1
-
-            argvals = []
-            for arg, m in zip(self.arg_spec, ms):
-                argvals.append(arg.values(m))
-
-            argarr = np.array(np.broadcast_arrays(*np.ix_(*argvals))).reshape(num_args, -1).T
-
-        # Check
-        old_dps, old_prec = mpmath.mp.dps, mpmath.mp.prec
-        try:
-            if self.dps is not None:
-                dps_list = [self.dps]
-            else:
-                dps_list = [20]
-            if self.prec is not None:
-                mpmath.mp.prec = self.prec
-
-            # Proper casting of mpmath input and output types. Using
-            # native mpmath types as inputs gives improved precision
-            # in some cases.
-            if np.issubdtype(argarr.dtype, np.complexfloating):
-                pytype = complex
-
-                def mptype(x):
-                    return mpmath.mpc(complex(x))
-            else:
-                def mptype(x):
-                    return mpmath.mpf(float(x))
-
-                def pytype(x):
-                    if abs(x.imag) > 1e-16*(1 + abs(x.real)):
-                        return np.nan
-                    else:
-                        return float(x.real)
-
-            # Try out different dps until one (or none) works
-            for j, dps in enumerate(dps_list):
-                mpmath.mp.dps = dps
-
-                try:
-                    assert_func_equal(self.scipy_func,
-                                      lambda *a: pytype(self.mpmath_func(*map(mptype, a))),
-                                      argarr,
-                                      vectorized=False,
-                                      rtol=self.rtol, atol=self.atol,
-                                      ignore_inf_sign=self.ignore_inf_sign,
-                                      nan_ok=self.nan_ok,
-                                      param_filter=self.param_filter)
-                    break
-                except AssertionError:
-                    if j >= len(dps_list)-1:
-                        reraise(*sys.exc_info())
-        finally:
-            mpmath.mp.dps, mpmath.mp.prec = old_dps, old_prec
-
-    def __repr__(self):
-        if self.is_complex:
-            return "<MpmathData: %s (complex)>" % (self.name,)
-        else:
-            return "<MpmathData: %s>" % (self.name,)
-
-
-def assert_mpmath_equal(*a, **kw):
-    d = MpmathData(*a, **kw)
-    d.check()
-
-
-def nonfunctional_tooslow(func):
-    return dec.skipif(True, "    Test not yet functional (too slow), needs more work.")(func)
-
-
-# ------------------------------------------------------------------------------
-# Dealing with mpmath quirks
-# ------------------------------------------------------------------------------
-
-def _trace_args(func):
-    def tofloat(x):
-        if isinstance(x, mpmath.mpc):
-            return complex(x)
-        else:
-            return float(x)
-
-    def wrap(*a, **kw):
-        sys.stderr.write("%r: " % (tuple(map(tofloat, a)),))
-        sys.stderr.flush()
-        try:
-            r = func(*a, **kw)
-            sys.stderr.write("-> %r" % r)
-        finally:
-            sys.stderr.write("\n")
-            sys.stderr.flush()
-        return r
-    return wrap
-
-try:
-    import posix
-    import signal
-    POSIX = ('setitimer' in dir(signal))
-except ImportError:
-    POSIX = False
-
-
-class _TimeoutError(Exception):
-    pass
-
-
-def _time_limited(timeout=0.5, return_val=np.nan, use_sigalrm=True):
-    """
-    Decorator for setting a timeout for pure-Python functions.
-
-    If the function does not return within `timeout` seconds, the
-    value `return_val` is returned instead.
-
-    On POSIX this uses SIGALRM by default. On non-POSIX, settrace is
-    used. Do not use this with threads: the SIGALRM implementation
-    does probably not work well. The settrace implementation only
-    traces the current thread.
-
-    The settrace implementation slows down execution speed. Slowdown
-    by a factor around 10 is probably typical.
-    """
-    if POSIX and use_sigalrm:
-        def sigalrm_handler(signum, frame):
-            raise _TimeoutError()
-
-        def deco(func):
-            def wrap(*a, **kw):
-                old_handler = signal.signal(signal.SIGALRM, sigalrm_handler)
-                signal.setitimer(signal.ITIMER_REAL, timeout)
-                try:
-                    return func(*a, **kw)
-                except _TimeoutError:
-                    return return_val
-                finally:
-                    signal.setitimer(signal.ITIMER_REAL, 0)
-                    signal.signal(signal.SIGALRM, old_handler)
-            return wrap
-    else:
-        def deco(func):
-            def wrap(*a, **kw):
-                start_time = time.time()
-
-                def trace(frame, event, arg):
-                    if time.time() - start_time > timeout:
-                        raise _TimeoutError()
-                    return None  # turn off tracing except at function calls
-                sys.settrace(trace)
-                try:
-                    return func(*a, **kw)
-                except _TimeoutError:
-                    sys.settrace(None)
-                    return return_val
-                finally:
-                    sys.settrace(None)
-            return wrap
-    return deco
-
-
-def _exception_to_nan(func):
-    """Decorate function to return nan if it raises an exception"""
-    def wrap(*a, **kw):
-        try:
-            return func(*a, **kw)
-        except Exception:
-            return np.nan
-    return wrap
-
-
-def _inf_to_nan(func):
-    """Decorate function to return nan if it returns inf"""
-    def wrap(*a, **kw):
-        v = func(*a, **kw)
-        if not np.isfinite(v):
-            return np.nan
-        return v
-    return wrap
 
 
 # ------------------------------------------------------------------------------
@@ -917,12 +608,12 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
 
     def test_bei(self):
         assert_mpmath_equal(sc.bei,
-                            _exception_to_nan(lambda z: mpmath.bei(0, z, **HYPERKW)),
+                            exception_to_nan(lambda z: mpmath.bei(0, z, **HYPERKW)),
                             [Arg(-1e3, 1e3)])
 
     def test_ber(self):
         assert_mpmath_equal(sc.ber,
-                            _exception_to_nan(lambda z: mpmath.ber(0, z, **HYPERKW)),
+                            exception_to_nan(lambda z: mpmath.ber(0, z, **HYPERKW)),
                             [Arg(-1e3, 1e3)])
 
     def test_bernoulli(self):
@@ -933,31 +624,31 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
 
     def test_besseli(self):
         assert_mpmath_equal(sc.iv,
-                            _exception_to_nan(lambda v, z: mpmath.besseli(v, z, **HYPERKW)),
+                            exception_to_nan(lambda v, z: mpmath.besseli(v, z, **HYPERKW)),
                             [Arg(-1e100, 1e100), Arg()],
                             atol=1e-270)
 
     def test_besseli_complex(self):
         assert_mpmath_equal(lambda v, z: sc.iv(v.real, z),
-                            _exception_to_nan(lambda v, z: mpmath.besseli(v, z, **HYPERKW)),
+                            exception_to_nan(lambda v, z: mpmath.besseli(v, z, **HYPERKW)),
                             [Arg(-1e100, 1e100), ComplexArg()])
 
     def test_besselj(self):
         assert_mpmath_equal(sc.jv,
-                            _exception_to_nan(lambda v, z: mpmath.besselj(v, z, **HYPERKW)),
+                            exception_to_nan(lambda v, z: mpmath.besselj(v, z, **HYPERKW)),
                             [Arg(-1e100, 1e100), Arg(-1e3, 1e3)],
                             ignore_inf_sign=True)
 
         # loss of precision at large arguments due to oscillation
         assert_mpmath_equal(sc.jv,
-                            _exception_to_nan(lambda v, z: mpmath.besselj(v, z, **HYPERKW)),
+                            exception_to_nan(lambda v, z: mpmath.besselj(v, z, **HYPERKW)),
                             [Arg(-1e100, 1e100), Arg(-1e8, 1e8)],
                             ignore_inf_sign=True,
                             rtol=1e-5)
 
     def test_besselj_complex(self):
         assert_mpmath_equal(lambda v, z: sc.jv(v.real, z),
-                            _exception_to_nan(lambda v, z: mpmath.besselj(v, z, **HYPERKW)),
+                            exception_to_nan(lambda v, z: mpmath.besselj(v, z, **HYPERKW)),
                             [Arg(), ComplexArg()])
 
     def test_besselk(self):
@@ -977,17 +668,17 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                     mpmath.mp.dps = old_dps
             return r
         assert_mpmath_equal(sc.kv,
-                            _exception_to_nan(mpbesselk),
+                            exception_to_nan(mpbesselk),
                             [Arg(-1e100, 1e100), Arg()])
 
     def test_besselk_int(self):
         assert_mpmath_equal(sc.kn,
-                            _exception_to_nan(lambda v, z: mpmath.besselk(v, z, **HYPERKW)),
+                            exception_to_nan(lambda v, z: mpmath.besselk(v, z, **HYPERKW)),
                             [IntArg(-1000, 1000), Arg()])
 
     def test_besselk_complex(self):
         assert_mpmath_equal(lambda v, z: sc.kv(v.real, z),
-                            _exception_to_nan(lambda v, z: mpmath.besselk(v, z, **HYPERKW)),
+                            exception_to_nan(lambda v, z: mpmath.besselk(v, z, **HYPERKW)),
                             [Arg(-1e100, 1e100), ComplexArg()])
 
     def test_bessely(self):
@@ -1001,7 +692,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return np.nan
             return r
         assert_mpmath_equal(sc.yv,
-                            _exception_to_nan(mpbessely),
+                            exception_to_nan(mpbessely),
                             [Arg(-1e100, 1e100), Arg(-1e8, 1e8)],
                             n=5000)
 
@@ -1017,7 +708,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                     np.seterr(**olderr)
             return r
         assert_mpmath_equal(lambda v, z: sc.yv(v.real, z),
-                            _exception_to_nan(mpbessely),
+                            exception_to_nan(mpbessely),
                             [Arg(), ComplexArg()],
                             n=15000)
 
@@ -1029,7 +720,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return np.nan
             return r
         assert_mpmath_equal(lambda v, z: sc.yn(int(v), z),
-                            _exception_to_nan(mpbessely),
+                            exception_to_nan(mpbessely),
                             [IntArg(-1000, 1000), Arg(-1e8, 1e8)])
 
     def test_beta(self):
@@ -1065,7 +756,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
 
     def test_betainc(self):
         assert_mpmath_equal(sc.betainc,
-                            _time_limited()(_exception_to_nan(lambda a, b, x: mpmath.betainc(a, b, 0, x, regularized=True))),
+                            time_limited()(exception_to_nan(lambda a, b, x: mpmath.betainc(a, b, 0, x, regularized=True))),
                             [Arg(), Arg(), Arg()])
 
     def test_binom(self):
@@ -1099,24 +790,24 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
 
     def test_chebyt_int(self):
         assert_mpmath_equal(lambda n, x: sc.eval_chebyt(int(n), x),
-                            _exception_to_nan(lambda n, x: mpmath.chebyt(n, x, **HYPERKW)),
+                            exception_to_nan(lambda n, x: mpmath.chebyt(n, x, **HYPERKW)),
                             [IntArg(), Arg()], dps=50)
 
     @knownfailure_overridable("some cases in hyp2f1 not fully accurate")
     def test_chebyt(self):
         assert_mpmath_equal(sc.eval_chebyt,
-                            lambda n, x: _time_limited()(_exception_to_nan(mpmath.chebyt))(n, x, **HYPERKW),
+                            lambda n, x: time_limited()(exception_to_nan(mpmath.chebyt))(n, x, **HYPERKW),
                             [Arg(-101, 101), Arg()], n=10000)
 
     def test_chebyu_int(self):
         assert_mpmath_equal(lambda n, x: sc.eval_chebyu(int(n), x),
-                            _exception_to_nan(lambda n, x: mpmath.chebyu(n, x, **HYPERKW)),
+                            exception_to_nan(lambda n, x: mpmath.chebyu(n, x, **HYPERKW)),
                             [IntArg(), Arg()], dps=50)
 
     @knownfailure_overridable("some cases in hyp2f1 not fully accurate")
     def test_chebyu(self):
         assert_mpmath_equal(sc.eval_chebyu,
-                            lambda n, x: _time_limited()(_exception_to_nan(mpmath.chebyu))(n, x, **HYPERKW),
+                            lambda n, x: time_limited()(exception_to_nan(mpmath.chebyu))(n, x, **HYPERKW),
                             [Arg(-101, 101), Arg()])
 
     def test_chi(self):
@@ -1149,7 +840,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
 
     def test_digamma(self):
         assert_mpmath_equal(sc.digamma,
-                            _exception_to_nan(mpmath.digamma),
+                            exception_to_nan(mpmath.digamma),
                             [Arg()], rtol=1e-12, dps=50)
 
     def test_digamma_complex(self):
@@ -1159,7 +850,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
             return np.where((z.real < 0) & (np.abs(z.imag) < 1.12), False, True)
 
         assert_mpmath_equal(sc.digamma,
-                            _exception_to_nan(mpmath.digamma),
+                            exception_to_nan(mpmath.digamma),
                             [ComplexArg()], rtol=1e-13, dps=40,
                             param_filter=param_filter)
 
@@ -1305,12 +996,12 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
 
     def test_erfc(self):
         assert_mpmath_equal(sc.erfc,
-                            _exception_to_nan(lambda z: mpmath.erfc(z)),
+                            exception_to_nan(lambda z: mpmath.erfc(z)),
                             [Arg()])
 
     def test_erfc_complex(self):
         assert_mpmath_equal(sc.erfc,
-                            _exception_to_nan(lambda z: mpmath.erfc(z)),
+                            exception_to_nan(lambda z: mpmath.erfc(z)),
                             [ComplexArg()], n=200)
 
     def test_erfi(self):
@@ -1325,7 +1016,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
 
     def test_ndtr(self):
         assert_mpmath_equal(sc.ndtr,
-                            _exception_to_nan(lambda z: mpmath.ncdf(z)),
+                            exception_to_nan(lambda z: mpmath.ncdf(z)),
                             [Arg()], n=200)
 
     def test_ndtr_complex(self):
@@ -1335,12 +1026,12 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
 
     def test_log_ndtr(self):
         assert_mpmath_equal(sc.log_ndtr,
-                            _exception_to_nan(lambda z: mpmath.log(mpmath.ncdf(z))),
+                            exception_to_nan(lambda z: mpmath.log(mpmath.ncdf(z))),
                             [Arg()], n=600, dps=300)
 
     def test_log_ndtr_complex(self):
         assert_mpmath_equal(sc.log_ndtr,
-                            _exception_to_nan(lambda z: mpmath.log(mpmath.erfc(-z/np.sqrt(2.))/2.)),
+                            exception_to_nan(lambda z: mpmath.log(mpmath.erfc(-z/np.sqrt(2.))/2.)),
                             [ComplexArg(a=complex(-10000, -100),
                                         b=complex(10000, 100))], n=200, dps=300)
 
@@ -1352,7 +1043,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
     @knownfailure_overridable("spurious(?) inf for negative x")
     def test_expint(self):
         assert_mpmath_equal(sc.expn,
-                            _exception_to_nan(mpmath.expint),
+                            exception_to_nan(mpmath.expint),
                             [IntArg(0, 100), Arg()])
 
     def test_fresnels(self):
@@ -1371,46 +1062,40 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
 
     def test_gamma(self):
         assert_mpmath_equal(sc.gamma,
-                            _exception_to_nan(mpmath.gamma),
+                            exception_to_nan(mpmath.gamma),
                             [Arg()])
 
     def test_gamma_complex(self):
         assert_mpmath_equal(sc.gamma,
-                            _exception_to_nan(mpmath.gamma),
+                            exception_to_nan(mpmath.gamma),
                             [ComplexArg()], rtol=1e-12)
 
     def test_gammainc(self):
-        # Mpmath uses hypergeometric functions to evaluate gammainc
-        # and (as of version 0.19) does not provided a keyword to
-        # increase the number of terms used. As a result it raises
-        # exceptions on large arguments due to lack of
-        # convergence. Since the _exception_to_nan wrapper would just
-        # ignore bad values, we first test in the region where there
-        # are no exceptions.
+        # Larger arguments are tested in test_data.py:test_local
         assert_mpmath_equal(sc.gammainc,
                             lambda z, b: mpmath.gammainc(z, b=b, regularized=True),
                             [Arg(0, 1e4, inclusive_a=False), Arg(0, 1e4)],
                             nan_ok=False, rtol=1e-11)
-        # Now add _exception_to_nan and go for broke. We still have to
-        # limit the range so that we don't have to use a crazy amount
-        # of precision.
-        assert_mpmath_equal(sc.gammainc,
-                            _exception_to_nan(
-                            lambda z, b: mpmath.gammainc(z, b=b, regularized=True)),
-                            [Arg(1e4, 1e100, inclusive_a=False), Arg(1e4, 1e100)],
-                            dps=100, rtol=1e-11)
+
+    @knownfailure_overridable()
+    def test_gammaincc(self):
+        # Larger arguments are tested in test_data.py:test_local
+        assert_mpmath_equal(sc.gammaincc,
+                            lambda z, a: mpmath.gammainc(z, a=a, regularized=True),
+                            [Arg(0, 1e4, inclusive_a=False), Arg(0, 1e4)],
+                             nan_ok=False, rtol=1e-10)
 
     def test_gammaln(self):
         # The real part of loggamma is log(|gamma(z)|).
         def f(z):
             return mpmath.loggamma(z).real
 
-        assert_mpmath_equal(sc.gammaln, _exception_to_nan(f), [Arg()])
+        assert_mpmath_equal(sc.gammaln, exception_to_nan(f), [Arg()])
 
     @knownfailure_overridable()
     def test_gegenbauer(self):
         assert_mpmath_equal(sc.eval_gegenbauer,
-                            _exception_to_nan(mpmath.gegenbauer),
+                            exception_to_nan(mpmath.gegenbauer),
                             [Arg(-1e3, 1e3), Arg(), Arg()])
 
     def test_gegenbauer_int(self):
@@ -1447,14 +1132,14 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return np.inf
             return r
         assert_mpmath_equal(sc_gegenbauer,
-                            _exception_to_nan(gegenbauer),
+                            exception_to_nan(gegenbauer),
                             [IntArg(0, 100), Arg(), Arg()],
                             n=40000, dps=100,
                             ignore_inf_sign=True)
 
         # Check the small-x expansion
         assert_mpmath_equal(sc_gegenbauer,
-                            _exception_to_nan(gegenbauer),
+                            exception_to_nan(gegenbauer),
                             [IntArg(0, 100), Arg(), FixedArg(np.logspace(-30, -4, 30))],
                             dps=100,
                             ignore_inf_sign=True)
@@ -1462,30 +1147,30 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
     @knownfailure_overridable()
     def test_gegenbauer_complex(self):
         assert_mpmath_equal(lambda n, a, x: sc.eval_gegenbauer(int(n), a.real, x),
-                            _exception_to_nan(mpmath.gegenbauer),
+                            exception_to_nan(mpmath.gegenbauer),
                             [IntArg(0, 100), Arg(), ComplexArg()])
 
     @nonfunctional_tooslow
     def test_gegenbauer_complex_general(self):
         assert_mpmath_equal(lambda n, a, x: sc.eval_gegenbauer(n.real, a.real, x),
-                            _exception_to_nan(mpmath.gegenbauer),
+                            exception_to_nan(mpmath.gegenbauer),
                             [Arg(-1e3, 1e3), Arg(), ComplexArg()])
 
     def test_hankel1(self):
         assert_mpmath_equal(sc.hankel1,
-                            _exception_to_nan(lambda v, x: mpmath.hankel1(v, x,
+                            exception_to_nan(lambda v, x: mpmath.hankel1(v, x,
                                                                           **HYPERKW)),
                             [Arg(-1e20, 1e20), Arg()])
 
     def test_hankel2(self):
         assert_mpmath_equal(sc.hankel2,
-                            _exception_to_nan(lambda v, x: mpmath.hankel2(v, x, **HYPERKW)),
+                            exception_to_nan(lambda v, x: mpmath.hankel2(v, x, **HYPERKW)),
                             [Arg(-1e20, 1e20), Arg()])
 
     @knownfailure_overridable("issues at intermediately large orders")
     def test_hermite(self):
         assert_mpmath_equal(lambda n, x: sc.eval_hermite(int(n), x),
-                            _exception_to_nan(mpmath.hermite),
+                            exception_to_nan(mpmath.hermite),
                             [IntArg(0, 10000), Arg()])
 
     # hurwitz: same as zeta
@@ -1504,7 +1189,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
     def test_hyp0f1_complex(self):
         KW = dict(maxprec=400, maxterms=1500)
         assert_mpmath_equal(lambda a, z: sc.hyp0f1(a.real, z),
-                            _exception_to_nan(lambda a, x: mpmath.hyp0f1(a, x, **HYPERKW)),
+                            exception_to_nan(lambda a, x: mpmath.hyp0f1(a, x, **HYPERKW)),
                             [Arg(-25, 25), ComplexArg(complex(-120, -120), complex(120, 120))])
         # NB: The range of the first parameter ("v") are limited by an overflow
         # in the intermediate calculations. Can be fixed by implementing an
@@ -1512,15 +1197,15 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
 
     @knownfailure_overridable()
     def test_hyp1f1(self):
-        assert_mpmath_equal(_inf_to_nan(sc.hyp1f1),
-                            _exception_to_nan(lambda a, b, x: mpmath.hyp1f1(a, b, x, **HYPERKW)),
+        assert_mpmath_equal(inf_to_nan(sc.hyp1f1),
+                            exception_to_nan(lambda a, b, x: mpmath.hyp1f1(a, b, x, **HYPERKW)),
                             [Arg(-1e5, 1e5), Arg(-1e5, 1e5), Arg()],
                             n=2000)
 
     @knownfailure_overridable()
     def test_hyp1f1_complex(self):
-        assert_mpmath_equal(_inf_to_nan(lambda a, b, x: sc.hyp1f1(a.real, b.real, x)),
-                            _exception_to_nan(lambda a, b, x: mpmath.hyp1f1(a, b, x, **HYPERKW)),
+        assert_mpmath_equal(inf_to_nan(lambda a, b, x: sc.hyp1f1(a.real, b.real, x)),
+                            exception_to_nan(lambda a, b, x: mpmath.hyp1f1(a, b, x, **HYPERKW)),
                             [Arg(-1e3, 1e3), Arg(-1e3, 1e3), ComplexArg()],
                             n=2000)
 
@@ -1532,7 +1217,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return np.nan
             return v
         assert_mpmath_equal(hyp1f2,
-                            _exception_to_nan(lambda a, b, c, x: mpmath.hyp1f2(a, b, c, x, **HYPERKW)),
+                            exception_to_nan(lambda a, b, c, x: mpmath.hyp1f2(a, b, c, x, **HYPERKW)),
                             [Arg(), Arg(), Arg(), Arg()],
                             n=20000)
 
@@ -1544,28 +1229,28 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return np.nan
             return v
         assert_mpmath_equal(hyp2f0,
-                            lambda a, b, x: _time_limited(0.1)(_exception_to_nan(_trace_args(mpmath.hyp2f0)))(
+                            lambda a, b, x: time_limited(0.1)(exception_to_nan(trace_args(mpmath.hyp2f0)))(
                                 a, b, x, **HYPERKW),
                             [Arg(), Arg(), Arg()])
 
     @knownfailure_overridable("spurious inf (or inf with wrong sign) for some argument values")
     def test_hyp2f1(self):
         assert_mpmath_equal(sc.hyp2f1,
-                            _exception_to_nan(lambda a, b, c, x: mpmath.hyp2f1(a, b, c, x, **HYPERKW)),
+                            exception_to_nan(lambda a, b, c, x: mpmath.hyp2f1(a, b, c, x, **HYPERKW)),
                             [Arg(), Arg(), Arg(), Arg()])
 
     @nonfunctional_tooslow
     def test_hyp2f1_complex(self):
         # Scipy's hyp2f1 seems to have performance and accuracy problems
         assert_mpmath_equal(lambda a, b, c, x: sc.hyp2f1(a.real, b.real, c.real, x),
-                            _exception_to_nan(lambda a, b, c, x: mpmath.hyp2f1(a, b, c, x, **HYPERKW)),
+                            exception_to_nan(lambda a, b, c, x: mpmath.hyp2f1(a, b, c, x, **HYPERKW)),
                             [Arg(-1e2, 1e2), Arg(-1e2, 1e2), Arg(-1e2, 1e2), ComplexArg()],
                             n=10)
 
     @knownfailure_overridable()
     def test_hyperu(self):
         assert_mpmath_equal(sc.hyperu,
-                            _exception_to_nan(lambda a, b, x: mpmath.hyperu(a, b, x, **HYPERKW)),
+                            exception_to_nan(lambda a, b, x: mpmath.hyperu(a, b, x, **HYPERKW)),
                             [Arg(), Arg(), Arg()])
 
     def test_j0(self):
@@ -1595,10 +1280,10 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
     @knownfailure_overridable()
     def test_jacobi(self):
         assert_mpmath_equal(sc.eval_jacobi,
-                            _exception_to_nan(lambda a, b, c, x: mpmath.jacobi(a, b, c, x, **HYPERKW)),
+                            exception_to_nan(lambda a, b, c, x: mpmath.jacobi(a, b, c, x, **HYPERKW)),
                             [Arg(), Arg(), Arg(), Arg()])
         assert_mpmath_equal(lambda n, b, c, x: sc.eval_jacobi(int(n), b, c, x),
-                            _exception_to_nan(lambda a, b, c, x: mpmath.jacobi(a, b, c, x, **HYPERKW)),
+                            exception_to_nan(lambda a, b, c, x: mpmath.jacobi(a, b, c, x, **HYPERKW)),
                             [IntArg(), Arg(), Arg(), Arg()])
 
     def test_jacobi_int(self):
@@ -1609,7 +1294,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return 1.0
             return mpmath.jacobi(n, a, b, x)
         assert_mpmath_equal(lambda n, a, b, x: sc.eval_jacobi(int(n), a, b, x),
-                            lambda n, a, b, x: _exception_to_nan(jacobi)(n, a, b, x, **HYPERKW),
+                            lambda n, a, b, x: exception_to_nan(jacobi)(n, a, b, x, **HYPERKW),
                             [IntArg(), Arg(), Arg(), Arg()],
                             n=20000, dps=50)
 
@@ -1618,25 +1303,25 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
             if x == 0:
                 # work around mpmath issue at x=0
                 return -pi/4
-            return _exception_to_nan(mpmath.kei)(0, x, **HYPERKW)
+            return exception_to_nan(mpmath.kei)(0, x, **HYPERKW)
         assert_mpmath_equal(sc.kei,
                             kei,
                             [Arg(-1e30, 1e30)], n=1000)
 
     def test_ker(self):
         assert_mpmath_equal(sc.ker,
-                            _exception_to_nan(lambda x: mpmath.ker(0, x, **HYPERKW)),
+                            exception_to_nan(lambda x: mpmath.ker(0, x, **HYPERKW)),
                             [Arg(-1e30, 1e30)], n=1000)
 
     @nonfunctional_tooslow
     def test_laguerre(self):
-        assert_mpmath_equal(_trace_args(sc.eval_laguerre),
-                            lambda n, x: _exception_to_nan(mpmath.laguerre)(n, x, **HYPERKW),
+        assert_mpmath_equal(trace_args(sc.eval_laguerre),
+                            lambda n, x: exception_to_nan(mpmath.laguerre)(n, x, **HYPERKW),
                             [Arg(), Arg()])
 
     def test_laguerre_int(self):
         assert_mpmath_equal(lambda n, x: sc.eval_laguerre(int(n), x),
-                            lambda n, x: _exception_to_nan(mpmath.laguerre)(n, x, **HYPERKW),
+                            lambda n, x: exception_to_nan(mpmath.laguerre)(n, x, **HYPERKW),
                             [IntArg(), Arg()], n=20000)
 
     def test_lambertw(self):
@@ -1652,13 +1337,13 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
 
     def test_legendre_int(self):
         assert_mpmath_equal(lambda n, x: sc.eval_legendre(int(n), x),
-                            lambda n, x: _exception_to_nan(mpmath.legendre)(n, x, **HYPERKW),
+                            lambda n, x: exception_to_nan(mpmath.legendre)(n, x, **HYPERKW),
                             [IntArg(), Arg()],
                             n=20000)
 
         # Check the small-x expansion
         assert_mpmath_equal(lambda n, x: sc.eval_legendre(int(n), x),
-                            lambda n, x: _exception_to_nan(mpmath.legendre)(n, x, **HYPERKW),
+                            lambda n, x: exception_to_nan(mpmath.legendre)(n, x, **HYPERKW),
                             [IntArg(), FixedArg(np.logspace(-30, -4, 20))])
 
     def test_legenp(self):
@@ -1695,7 +1380,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return np.nan
 
             typ = 2 if abs(z) < 1 else 3
-            v = _exception_to_nan(mpmath.legenp)(n, m, z, type=typ)
+            v = exception_to_nan(mpmath.legenp)(n, m, z, type=typ)
 
             if abs(v) > 1e306:
                 # harmonize overflow to inf
@@ -1722,7 +1407,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
             if abs(z) < 1e-15:
                 # mpmath has bad performance here
                 return np.nan
-            return _exception_to_nan(mpmath.legenp)(int(n.real), int(m.real), z, type=2)
+            return exception_to_nan(mpmath.legenp)(int(n.real), int(m.real), z, type=2)
 
         # mpmath is quite slow here
         x = np.array([-2, -0.99, -0.5, 0, 1e-5, 0.5, 0.99, 20, 2e3])
@@ -1746,7 +1431,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
             if abs(z) < 1e-15:
                 # mpmath has bad performance here
                 return np.nan
-            return _exception_to_nan(mpmath.legenp)(int(n.real), int(m.real), z, type=3)
+            return exception_to_nan(mpmath.legenp)(int(n.real), int(m.real), z, type=3)
 
         # mpmath is quite slow here
         x = np.array([-2, -0.99, -0.5, 0, 1e-5, 0.5, 0.99, 20, 2e3])
@@ -1768,7 +1453,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
             if abs(z) < 1e-15:
                 # mpmath has bad performance here
                 return np.nan
-            return _exception_to_nan(mpmath.legenq)(n, m, z, type=2)
+            return exception_to_nan(mpmath.legenq)(n, m, z, type=2)
 
         assert_mpmath_equal(lqnm,
                             legenq,
@@ -1783,7 +1468,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
             if abs(z) < 1e-15:
                 # mpmath has bad performance here
                 return np.nan
-            return _exception_to_nan(mpmath.legenq)(int(n.real), int(m.real), z, type=2)
+            return exception_to_nan(mpmath.legenq)(int(n.real), int(m.real), z, type=2)
 
         assert_mpmath_equal(lqnm,
                             legenq,
@@ -1792,12 +1477,12 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
 
     def test_loggamma(self):
         assert_mpmath_equal(sc.loggamma,
-                            _exception_to_nan(lambda x: mpmath.loggamma(x, **HYPERKW)),
+                            exception_to_nan(lambda x: mpmath.loggamma(x, **HYPERKW)),
                             [Arg()], rtol=1e-13)
 
     def test_loggamma_complex(self):
         assert_mpmath_equal(sc.loggamma,
-                            _exception_to_nan(lambda z: mpmath.loggamma(z, **HYPERKW)),
+                            exception_to_nan(lambda z: mpmath.loggamma(z, **HYPERKW)),
                             [ComplexArg()], rtol=1e-13)
 
     @knownfailure_overridable()
@@ -1805,7 +1490,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
         def pcfd(v, x):
             return sc.pbdv(v, x)[0]
         assert_mpmath_equal(pcfd,
-                            _exception_to_nan(lambda v, x: mpmath.pcfd(v, x, **HYPERKW)),
+                            exception_to_nan(lambda v, x: mpmath.pcfd(v, x, **HYPERKW)),
                             [Arg(), Arg()])
 
     @knownfailure_overridable("it's not the same as the mpmath function --- maybe different definition?")
@@ -1813,7 +1498,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
         def pcfv(v, x):
             return sc.pbvv(v, x)[0]
         assert_mpmath_equal(pcfv,
-                            lambda v, x: _time_limited()(_exception_to_nan(mpmath.pcfv))(v, x, **HYPERKW),
+                            lambda v, x: time_limited()(exception_to_nan(mpmath.pcfv))(v, x, **HYPERKW),
                             [Arg(), Arg()], n=1000)
 
     @knownfailure_overridable()
@@ -1821,13 +1506,13 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
         def pcfw(a, x):
             return sc.pbwa(a, x)[0]
         assert_mpmath_equal(pcfw,
-                            lambda v, x: _time_limited()(_exception_to_nan(mpmath.pcfw))(v, x, **HYPERKW),
+                            lambda v, x: time_limited()(exception_to_nan(mpmath.pcfw))(v, x, **HYPERKW),
                             [Arg(), Arg()], dps=50, n=1000)
 
     @knownfailure_overridable("issues at large arguments (atol OK, rtol not) and <eps-close to z=0")
     def test_polygamma(self):
         assert_mpmath_equal(sc.polygamma,
-                            _time_limited()(_exception_to_nan(mpmath.polygamma)),
+                            time_limited()(exception_to_nan(mpmath.polygamma)),
                             [IntArg(0, 1000), Arg()])
 
     def test_rgamma(self):
@@ -1844,7 +1529,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
 
     def test_rgamma_complex(self):
         assert_mpmath_equal(sc.rgamma,
-                            _exception_to_nan(mpmath.rgamma),
+                            exception_to_nan(mpmath.rgamma),
                             [ComplexArg()], rtol=1e-12)
 
     def test_rf(self):
@@ -1889,14 +1574,14 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
             return mpmath.polylog(2, 1 - x)
         # Spence has a branch cut on the negative real axis
         assert_mpmath_equal(sc.spence,
-                            _exception_to_nan(dilog),
+                            exception_to_nan(dilog),
                             [Arg(0, np.inf)], rtol=1e-14)
 
     def test_spence_complex(self):
         def dilog(z):
             return mpmath.polylog(2, 1 - z)
         assert_mpmath_equal(sc.spence,
-                            _exception_to_nan(dilog),
+                            exception_to_nan(dilog),
                             [ComplexArg()], rtol=1e-14)
 
     def test_spherharm(self):
@@ -1913,7 +1598,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
 
     def test_struveh(self):
         assert_mpmath_equal(sc.struve,
-                            _exception_to_nan(mpmath.struveh),
+                            exception_to_nan(mpmath.struveh),
                             [Arg(-1e4, 1e4), Arg(0, 1e4)],
                             rtol=5e-10)
 
@@ -1930,14 +1615,14 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
             return mpmath.struvel(v, z)
 
         assert_mpmath_equal(sc.modstruve,
-                            _exception_to_nan(mp_struvel),
+                            exception_to_nan(mp_struvel),
                             [Arg(-1e4, 1e4), Arg(0, 1e4)],
                             rtol=5e-10,
                             ignore_inf_sign=True)
 
     def test_zeta(self):
         assert_mpmath_equal(sc.zeta,
-                            _exception_to_nan(mpmath.zeta),
+                            exception_to_nan(mpmath.zeta),
                             [Arg(a=1, b=1e10, inclusive_a=False),
                              Arg(a=0, inclusive_a=False)])
 
@@ -1952,7 +1637,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return mpmath.mp.powm1(x, lmbda) / lmbda
 
         assert_mpmath_equal(sc.boxcox,
-                            _exception_to_nan(mp_boxcox),
+                            exception_to_nan(mp_boxcox),
                             [Arg(a=0, inclusive_a=False), Arg()],
                             n=200,
                             dps=60,
@@ -1970,7 +1655,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return mpmath.mp.powm1(one + x, lmbda) / lmbda
 
         assert_mpmath_equal(sc.boxcox1p,
-                            _exception_to_nan(mp_boxcox1p),
+                            exception_to_nan(mp_boxcox1p),
                             [Arg(a=-1, inclusive_a=False), Arg()],
                             n=200,
                             dps=60,
@@ -1987,7 +1672,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return out
 
         assert_mpmath_equal(lambda n, z: sc.spherical_jn(int(n), z),
-                            _exception_to_nan(mp_spherical_jn),
+                            exception_to_nan(mp_spherical_jn),
                             [IntArg(0, 200), Arg(-1e8, 1e8)],
                             dps=300)
 
@@ -2002,7 +1687,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return out
 
         assert_mpmath_equal(lambda n, z: sc.spherical_jn(int(n.real), z),
-                            _exception_to_nan(mp_spherical_jn),
+                            exception_to_nan(mp_spherical_jn),
                             [IntArg(0, 200), ComplexArg()])
 
     def test_spherical_yn(self):
@@ -2016,7 +1701,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return out
 
         assert_mpmath_equal(lambda n, z: sc.spherical_yn(int(n), z),
-                            _exception_to_nan(mp_spherical_yn),
+                            exception_to_nan(mp_spherical_yn),
                             [IntArg(0, 200), Arg(-1e10, 1e10)],
                             dps=100)
 
@@ -2031,7 +1716,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return out
 
         assert_mpmath_equal(lambda n, z: sc.spherical_yn(int(n.real), z),
-                            _exception_to_nan(mp_spherical_yn),
+                            exception_to_nan(mp_spherical_yn),
                             [IntArg(0, 200), ComplexArg()])
 
     def test_spherical_in(self):
@@ -2045,7 +1730,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return out
 
         assert_mpmath_equal(lambda n, z: sc.spherical_in(int(n), z),
-                            _exception_to_nan(mp_spherical_in),
+                            exception_to_nan(mp_spherical_in),
                             [IntArg(0, 200), Arg()],
                             dps=200, atol=10**(-278))
 
@@ -2060,7 +1745,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return out
 
         assert_mpmath_equal(lambda n, z: sc.spherical_in(int(n.real), z),
-                            _exception_to_nan(mp_spherical_in),
+                            exception_to_nan(mp_spherical_in),
                             [IntArg(0, 200), ComplexArg()])
 
     def test_spherical_kn(self):
@@ -2073,7 +1758,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return out
 
         assert_mpmath_equal(lambda n, z: sc.spherical_kn(int(n), z),
-                            _exception_to_nan(mp_spherical_kn),
+                            exception_to_nan(mp_spherical_kn),
                             [IntArg(0, 150), Arg()],
                             dps=100)
 
@@ -2089,7 +1774,7 @@ class TestSystematic(with_metaclass(DecoratorMeta, object)):
                 return out
 
         assert_mpmath_equal(lambda n, z: sc.spherical_kn(int(n.real), z),
-                            _exception_to_nan(mp_spherical_kn),
+                            exception_to_nan(mp_spherical_kn),
                             [IntArg(0, 200), ComplexArg()],
                             dps=200)
