+++ scipy/special/tests/test_precompute_utils.py
@@ -3,7 +3,8 @@ from scipy._lib.six import with_metaclass
 from numpy.testing import dec
 
 from scipy.special._testutils import MissingModule, check_version, DecoratorMeta
-from scipy.special._precompute import utils
+from scipy.special._mptestutils import mp_assert_allclose
+from scipy.special._precompute.utils import lagrange_inversion
 
 try:
     import sympy
@@ -28,12 +29,12 @@ class TestInversion(with_metaclass(DecoratorMeta, object)):
         with mp.workdps(30):
             logcoeffs = mp.taylor(lambda x: mp.log(1 + x), 0, 10)
             expcoeffs = mp.taylor(lambda x: mp.exp(x) - 1, 0, 10)
-            invlogcoeffs = utils.lagrange_inversion(logcoeffs)
-            utils.mpf_assert_allclose(invlogcoeffs, expcoeffs)
+            invlogcoeffs = lagrange_inversion(logcoeffs)
+            mp_assert_allclose(invlogcoeffs, expcoeffs)
 
     def test_sin(self):
         with mp.workdps(30):
             sincoeffs = mp.taylor(mp.sin, 0, 10)
             asincoeffs = mp.taylor(mp.asin, 0, 10)
-            invsincoeffs = utils.lagrange_inversion(sincoeffs)
-            utils.mpf_assert_allclose(invsincoeffs, asincoeffs, atol=1e-30)
+            invsincoeffs = lagrange_inversion(sincoeffs)
+            mp_assert_allclose(invsincoeffs, asincoeffs, atol=1e-30)
